requirements.txt
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
binance-connector==1.1.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.6
zope.interface==5.4.0
